requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib==3.3.0
numpy==1.19.4
opencv-python==4.6.0.66
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
# torch==1.9.0
# torchvision==0.10.0
tqdm>=4.41.0
protobuf<4.21.3
pydantic

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# fastapi -------------------------------------
fastapi==0.75.2
uvicorn==0.16.0
python-multipart
click==7.1.2
asgiref==3.4.1
gunicorn==20.1.0

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
# psutil  # system utilization
# thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
